Move out metadata and util package
