+++ sharding-scaling-mysql/src/main/java/info/avalon566/shardingscaling/mysql/MySQLBinlogReader.java
@@ -25,7 +25,7 @@ import info.avalon566.shardingscaling.core.execute.executor.reader.LogReader;
 import info.avalon566.shardingscaling.core.execute.executor.reader.NopLogPosition;
 import info.avalon566.shardingscaling.core.execute.executor.record.Column;
 import info.avalon566.shardingscaling.core.execute.executor.record.DataRecord;
-import info.avalon566.shardingscaling.core.execute.metadata.JdbcUri;
+import info.avalon566.shardingscaling.core.metadata.JdbcUri;
 import info.avalon566.shardingscaling.core.execute.executor.record.FinishedRecord;
 import info.avalon566.shardingscaling.core.execute.executor.record.PlaceholderRecord;
 import info.avalon566.shardingscaling.core.execute.executor.record.Record;
@@ -34,7 +34,7 @@ import info.avalon566.shardingscaling.mysql.binlog.event.DeleteRowsEvent;
 import info.avalon566.shardingscaling.mysql.binlog.event.PlaceholderEvent;
 import info.avalon566.shardingscaling.mysql.binlog.event.UpdateRowsEvent;
 import info.avalon566.shardingscaling.mysql.binlog.event.WriteRowsEvent;
-import info.avalon566.shardingscaling.core.execute.util.DataSourceFactory;
+import info.avalon566.shardingscaling.core.util.DataSourceFactory;
 import info.avalon566.shardingscaling.core.config.JdbcDataSourceConfiguration;
 import info.avalon566.shardingscaling.mysql.binlog.MySQLConnector;
 import lombok.Setter;
