+++ sharding-scaling-mysql/src/main/java/info/avalon566/shardingscaling/mysql/MySQLJdbcReader.java
@@ -20,7 +20,7 @@ package info.avalon566.shardingscaling.mysql;
 import info.avalon566.shardingscaling.core.config.JdbcDataSourceConfiguration;
 import info.avalon566.shardingscaling.core.config.RdbmsConfiguration;
 import info.avalon566.shardingscaling.core.execute.executor.reader.AbstractJdbcReader;
-import info.avalon566.shardingscaling.core.execute.metadata.JdbcUri;
+import info.avalon566.shardingscaling.core.metadata.JdbcUri;
 
 import java.util.List;
 import java.util.Map;
